refactor: move MerkleTree and ProofRequest classes to shared pkg
