+++ shared/src/index.ts
@@ -2,4 +2,7 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './ExcludableMerkleTree'
+export * from './MerkleTree'
+export * from './ProofRequest'
 export * from './helpers'
